requirements.txt
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
asgiref==3.7.2
attrs==23.2.0
binaryornot==0.4.4
blinker==1.7.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
copier==9.1.1
decorator==5.1.1
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
django-countries==7.5.1
django-rest-framework==0.1.0
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
dunamai==1.19.2
et-xmlfile==1.1.0
fastapi==0.110.0
filelock==3.13.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flet==0.21.1
flet-core==0.21.1
flet-runtime==0.21.1
funcy==2.0
greenlet==3.0.3
h11==0.14.0
holidays==0.44
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jinja2-ansible-filters==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
plumbum==1.8.2
prompt-toolkit==3.0.36
psycopg2==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyyaml-include==1.3.2
qrcode==7.4.2
questionary==2.0.1
referencing==0.33.0
repath==0.9.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
starlette==0.36.3
swagger-spec-validator==3.0.3
text-unidecode==1.3
types-python-dateutil==2.8.19.20240311
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.28.0
virtualenv==20.25.1
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
xlrd==2.0.1